requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
jinja2==3.1.4
python-multipart==0.0.20

SQLAlchemy==2.0.36
psycopg2-binary==2.9.10

python-jose==3.3.0
passlib[bcrypt]==1.7.4

pydantic==2.10.3
pydantic-settings==2.6.1

httpx==0.28.0
pytest==8.3.4
